web/portal-api/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.6

# Email (for invites)
aiosmtplib==3.0.1
email-validator==2.1.0

# API integrations
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
